Artifactory.cmake: Rewrite fetching code as a Python script
This is actually more code than the CMake equivalent, but hopefully
it is a lot more clear what is actually going on.

The behaviour is not identical. Changes include:

  - The initial check downloads all the POM files, rather than all the
    main artifacts. These should mean much less data is transferred when
    there are multiple compatible artifacts available.

  - Artifacts that don't have a hyphen in their version string will
    no longer be ignored (there was a bug in the CMake code that was
    usually not triggered).

  - The version and main artifact filename are worked out by reading
    the POM file, instead of by parsing the filenames. This is more
    robust than the filename parsing.

Due to a bug in the artifact-upload script, all existing artifacts will
be ignored, because the version in the POM file will be wrong.