artifactory-upload: Set correct 'version' in the POM file
We should be using the upload version (e.g. master-20160101-125125-4)
not the other version (master-SNAPSHOT) since the POM file is describing
that specific artifact.

The current code to fetch artifacts parses the filename to find the
version number, so this wasn't causing any problems, but I want to
switch to using version number from the POM file (because parsing
filenames always has the potential to break in weird ways).